Fixed some API problems. Remove one Repast dependecy.
When using the API in a test example, simulations using RepACL failed
to run due to a conflict between two copies of the repast JAR. 
Specifically, the reference to the Context Repast class in the DF was
the source of this conflict. By changing the type of the context field
in the DF to Context's interface "Collection<T>", I fixed this issue
while maintaining the functionality in RepACL the same.